Update Evolution to generate a better tests template.
Rename template-tests-UI.ts to common-UI-tests-helpers-template.
This will help me do tests more easily.
